Handle trillions of iterations in progress

+++ kelvinLocals.h
@@ -32,6 +32,6 @@
   int threadCount = 0;
 #endif
 
-  int cL[9] = { 0, 0, 0, 0, 0, 0, 0, 0, 0 },    ///< Actual # calls to each instance of compute_likelihood
+  unsigned long cL[9] = { 0, 0, 0, 0, 0, 0, 0, 0, 0 },    ///< Actual # calls to each instance of compute_likelihood
     eCL[9] = {
   0, 0, 0, 0, 0, 0, 0, 0, 0};   ///< Est. final # calls to each instance of compute_likelihood
